refactor(qimagpt): SP-14522 optimization
