+++ client/src/components/Auth/index.ts
@@ -1,5 +1,4 @@
-export { default as Login } from './Login';
-export { default as LoginQima } from './LoginQima';
+export { default as Login } from './LoginQima';
 export { default as Registration } from './Registration';
 export { default as ResetPassword } from './ResetPassword';
 export { default as ApiErrorWatcher } from './ApiErrorWatcher';
