+++ client/src/components/Endpoints/Settings/index.ts
@@ -1,5 +1,4 @@
-export { default as OpenAISettings } from './OpenAI';
-export { default as OpenAIQimaSettings } from './OpenAIQima';
+export { default as OpenAISettings } from './OpenAIQima';
 export { default as BingAISettings } from './BingAI';
 export { default as GoogleSettings } from './Google';
 export { default as PluginsSettings } from './Plugins';
