+++ client/src/localization/languages/It.tsx
@@ -7,11 +7,6 @@ export default {
   com_ui_example_10_year_old_b_day:
     'Hai qualche idea creativa per il compleanno di un bambino di 10 anni?',
   com_ui_example_http_in_js: 'Come posso fare una richiesta HTTP in Javascript?',
-  com_ui_example_inspection:
-    'Can you write a pitch of 3 sentences to explain to a client the difference between a pre-shipment inspection and an inline inspection?',
-  com_ui_example_food_email: 'Create an automated response for incoming food related emails.',
-  com_ui_example_report:
-    'Can you summarize in 5 bullet points the key insights from this skin research report.',
   com_ui_capabilities: 'Capacità',
   com_ui_capability_remember: 'Ricorda ciò che l\'utente ha detto in precedenza nella conversazione',
   com_ui_capability_correction: 'Permette all\'utente di fornire correzioni di seguito',
@@ -21,8 +16,6 @@ export default {
   com_ui_limitation_harmful_biased:
     'Può occasionalmente produrre istruzioni dannose o contenuti prevenuti',
   com_ui_limitation_limited_2021: 'Conoscenza limitata del mondo e degli eventi successivi al 2021',
-  com_ui_limitation_limited_qima:
-    'WARNING: Data is not trained on QIMA specific information and has limited knowledge of the world and events after 2021.',
   com_ui_input: 'Input',
   com_ui_close: 'Chiudi',
   com_ui_model: 'Modello',
@@ -110,7 +103,6 @@ export default {
   com_auth_to_try_again: 'per riprovare.',
   com_auth_submit_registration: 'Invia registrazione',
   com_auth_welcome_back: 'Bentornato',
-  com_auth_welcome_back_qima: 'Welcome back to QimaGPT',
   com_endpoint_open_menu: 'Apri menu',
   com_endpoint_bing_enable_sydney: 'Abilita Sydney',
   com_endpoint_bing_to_enable_sydney: 'Per abilitare Sydney',
