+++ client/src/localization/languages/Sv.tsx
@@ -6,11 +6,6 @@ export default {
   com_ui_example_quantum_computing: 'Förklara kvantberäkning på enkla termer', // Explain quantum computing in simple terms
   com_ui_example_10_year_old_b_day: 'Har du några kreativa idéer till ett 10-årings födelsedag?', // Got any creative ideas for a 10 year old's birthday?
   com_ui_example_http_in_js: 'Hur gör jag en HTTP-förfrågan i Javascript?', // How do I make an HTTP request in Javascript?
-  com_ui_example_inspection:
-    'Can you write a pitch of 3 sentences to explain to a client the difference between a pre-shipment inspection and an inline inspection?',
-  com_ui_example_food_email: 'Create an automated response for incoming food related emails.',
-  com_ui_example_report:
-    'Can you summarize in 5 bullet points the key insights from this skin research report.',
   com_ui_capabilities: 'Möjligheter', // Capabilities
   com_ui_capability_remember: 'Kommer ihåg vad användaren sa tidigare i konversationen', // Remembers what user said earlier in the conversation
   com_ui_capability_correction: 'Tillåter användaren att ge uppföljande rättelser', // Allows user to provide follow-up corrections
@@ -20,8 +15,6 @@ export default {
   com_ui_limitation_harmful_biased:
     'Kan ibland producera skadliga instruktioner eller snedvridet innehåll', // May occasionally produce harmful instructions or biased content
   com_ui_limitation_limited_2021: 'Begränsad kunskap om världen och händelser efter 2021', // Limited knowledge of world and events after 2021
-  com_ui_limitation_limited_qima:
-    'WARNING: Data is not trained on QIMA specific information and has limited knowledge of the world and events after 2021.',
   com_ui_input: 'Inmatning', // Input
   com_ui_close: 'Stäng', // Close
   com_ui_model: 'Modell', // Model
@@ -110,7 +103,6 @@ export default {
   com_auth_to_try_again: 'för att försöka igen.', // to try again.
   com_auth_submit_registration: 'Skicka registrering', // Submit registration
   com_auth_welcome_back: 'Välkommen tillbaka', // Welcome back
-  com_auth_welcome_back_qima: 'Welcome back to QimaGPT',
   com_endpoint_open_menu: 'Öppna meny', // Open Menu
   com_endpoint_bing_enable_sydney: 'Aktivera Sydney', // Enable Sydney
   com_endpoint_bing_to_enable_sydney: 'För att aktivera Sydney', // To enable Sydney
